+++ packages/autotool/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @llamaindex/autotool
 
+## 3.0.7
+
+### Patch Changes
+
+- Updated dependencies [23bcc37]
+  - llamaindex@0.6.7
+
+## 3.0.6
+
+### Patch Changes
+
+- Updated dependencies [d902cc3]
+- Updated dependencies [025ffe6]
+- Updated dependencies [a659574]
+  - llamaindex@0.6.6
+
 ## 3.0.5
 
 ### Patch Changes