+++ packages/autotool/examples/01_node/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @llamaindex/autotool-01-node-example
 
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies [23bcc37]
+  - llamaindex@0.6.7
+  - @llamaindex/autotool@3.0.7
+
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies [d902cc3]
+- Updated dependencies [025ffe6]
+- Updated dependencies [a659574]
+  - llamaindex@0.6.6
+  - @llamaindex/autotool@3.0.6
+
 ## 0.0.14
 
 ### Patch Changes