+++ packages/experimental/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @llamaindex/experimental
 
+## 0.0.85
+
+### Patch Changes
+
+- Updated dependencies [23bcc37]
+  - llamaindex@0.6.7
+
+## 0.0.84
+
+### Patch Changes
+
+- Updated dependencies [d902cc3]
+- Updated dependencies [025ffe6]
+- Updated dependencies [a659574]
+  - llamaindex@0.6.6
+
 ## 0.0.83
 
 ### Patch Changes