+++ packages/llamaindex/e2e/examples/cloudflare-worker-agent/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @llamaindex/cloudflare-worker-agent-test
 
+## 0.0.69
+
+### Patch Changes
+
+- Updated dependencies [23bcc37]
+  - llamaindex@0.6.7
+
+## 0.0.68
+
+### Patch Changes
+
+- Updated dependencies [d902cc3]
+- Updated dependencies [025ffe6]
+- Updated dependencies [a659574]
+  - llamaindex@0.6.6
+
 ## 0.0.67
 
 ### Patch Changes