+++ packages/llamaindex/e2e/examples/nextjs-node-runtime/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @llamaindex/next-node-runtime
 
+## 0.0.50
+
+### Patch Changes
+
+- Updated dependencies [23bcc37]
+  - llamaindex@0.6.7
+
+## 0.0.49
+
+### Patch Changes
+
+- Updated dependencies [d902cc3]
+- Updated dependencies [025ffe6]
+- Updated dependencies [a659574]
+  - llamaindex@0.6.6
+
 ## 0.0.48
 
 ### Patch Changes